Add working group information to website
This adds the charter and working group details to the published
website.

It is less than ideal at this point since it is duplicating the content
from the `wg` folder in order to make it available for hugo rendering.
The current tools like `readFile` from the docsy template does not allow
pulling in files from outside the website directory.

Until we can find a workaround for that, this just copies the content.
Once we find a better way, we should be able to keep the working group
charters as the single source of truth and just pull it in to the
rendered content. But for now, this makes a nice rendered view of our
working groups to make it easier for folks to find and learn about the
various working groups in TAG Runtime.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>